+++ NEWS.md
@@ -15,6 +15,7 @@
 * Reduce memory consumption by storing buds and branches as lightweight references when `memory` is `"transient"` (#1364).
 * Replace the `memory` class with the new `lookup` class.
 * Implement `memory = "auto"` to select transient memory for dynamic branches and persistent memory for other targets (#1371).
+* Omit whole pattern targets from branch subpipelines when possible. Should reduce memory consumption in some cases.
 
 # targets 1.8.0