+++ NEWS.md
@@ -16,6 +16,7 @@
 * Replace the `memory` class with the new `lookup` class.
 * Implement `memory = "auto"` to select transient memory for dynamic branches and persistent memory for other targets (#1371).
 * Omit whole pattern targets from branch subpipelines when possible. Should reduce memory consumption in some cases.
+* Omit whole stem targets from branch subpipelines when `retrieval` is `"main"` and only a bud is actually used.
 
 # targets 1.8.0